requirements.txt
pynput==1.7.6
python-nmap==0.7.1
six==1.16.0